+++ packages/internal/src/introspection/index.ts
@@ -1,0 +1,4 @@
+import type { IntrospectionQuery, IntrospectionType } from 'graphql';
+
+export * from './output';
+export { preprocessIntrospection } from './preprocess';
