+++ packages/internal/src/introspection/preprocess.ts
@@ -1,23 +1,7 @@
 import type { IntrospectionQuery, IntrospectionType } from 'graphql';
 
-import { minifyIntrospectionQuery } from '@urql/introspection';
-
-import { PREAMBLE_IGNORE, ANNOTATION_DTS, ANNOTATION_TS } from './constants';
-import { TadaError } from './errors';
-
-const stringifyJson = (input: unknown | string): string =>
-  typeof input === 'string' ? input : JSON.stringify(input, null, 2);
 const stringifyName = (input: string | undefined | null): string =>
   input ? `"${input}"` : 'never';
-
-export function minifyIntrospection(introspection: IntrospectionQuery): IntrospectionQuery {
-  return minifyIntrospectionQuery(introspection, {
-    includeDirectives: false,
-    includeEnums: true,
-    includeInputs: true,
-    includeScalars: true,
-  });
-}
 
 const stringifyDeepObjectType = (object: Record<string, unknown>) => {
   let output = '';
@@ -106,9 +90,7 @@ export const printIntrospectionType = (type: IntrospectionType) => {
   }
 };
 
-export async function preprocessIntrospection({
-  __schema: schema,
-}: IntrospectionQuery): Promise<string> {
+export function preprocessIntrospection({ __schema: schema }: IntrospectionQuery): string {
   const queryName = stringifyName(schema.queryType.name);
   const mutationName = stringifyName(schema.mutationType && schema.mutationType.name);
   const subscriptionName = stringifyName(schema.subscriptionType && schema.subscriptionType.name);
@@ -125,45 +107,5 @@ export async function preprocessIntrospection({
     `  mutation: ${mutationName};\n` +
     `  subscription: ${subscriptionName};\n` +
     `  types: {\n${evaluatedTypes}  };\n}`
-  );
-}
-
-interface OutputIntrospectionFileOptions {
-  fileType: '.ts' | '.d.ts' | string;
-  shouldPreprocess?: boolean;
-}
-
-export async function outputIntrospectionFile(
-  introspection: IntrospectionQuery | string,
-  opts: OutputIntrospectionFileOptions
-): Promise<string> {
-  if (/\.d\.ts$/.test(opts.fileType)) {
-    const json =
-      typeof introspection !== 'string' && opts.shouldPreprocess
-        ? await preprocessIntrospection(introspection)
-        : stringifyJson(introspection);
-    return [
-      PREAMBLE_IGNORE,
-      ANNOTATION_DTS,
-      `export type introspection = ${json};\n`,
-      "import * as gqlTada from 'gql.tada';\n",
-      "declare module 'gql.tada' {",
-      '  interface setupSchema {',
-      '    introspection: introspection',
-      '  }',
-      '}',
-    ].join('\n');
-  } else if (/\.ts$/.test(opts.fileType)) {
-    const json = stringifyJson(introspection);
-    return [
-      PREAMBLE_IGNORE,
-      ANNOTATION_TS,
-      `const introspection = ${json} as const;\n`,
-      'export { introspection };',
-    ].join('\n');
-  }
-
-  throw new TadaError(
-    `No available introspection format for "${opts.fileType}" (expected ".ts" or ".d.ts")`
   );
 }
